+++ order-service/src/main/java/com/rushcrew/order/infrastructure/dto/timedeal/StockReservationResponse.java
@@ -1,0 +1,14 @@
+package com.rushcrew.order.infrastructure.dto.timedeal;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+@Getter
+@NoArgsConstructor
+@AllArgsConstructor
+public class StockReservationResponse {
+	private boolean success;
+	private Integer availableStock;
+	private String message;
+}
