+++ order-service/src/main/java/com/rushcrew/order/infrastructure/dto/timedeal/TimeDealResponseStatus.java
@@ -1,0 +1,8 @@
+package com.rushcrew.order.infrastructure.dto.timedeal;
+
+public enum TimeDealResponseStatus {
+	SCHEDULED,
+	IN_PROGRESS,
+	SOLD_OUT,
+	ENDED
+}
